refactor(scope): extract resolveClassDefFromTypeMeta, resolveSuperclass, hasEngineClassConstant into ClassRegistry
- Remove 6 duplicate private method definitions across FrontendTopBindingAnalyzer,
  FrontendChainReductionHelper, and FrontendConstructorResolutionSupport
- Route all callers through the three new public ClassRegistry methods
